+++ examples/postman_echo/request_methods/demo_testsuite_yml/request_with_testcase_reference_test.py
@@ -29,7 +29,6 @@ class TestCaseRequestWithTestcaseReference(HttpRunner):
         )
         .base_url("https://postman-echo.com")
         .verify(False)
-        .locust_weight(3)
     )
 
     teststeps = [